Fix copying _out of_ IE
